modules/voice_cloning/misaki_vc.py
"""
Misaki Voice Cloning Implementation
"""

import asyncio
import logging
from typing import Dict, Any, Optional, List
from pathlib import Path

from .base import VoiceCloningBase


class MisakiVC(VoiceCloningBase):
    """Misaki voice cloning implementation"""
    
    def __init__(self):
        super().__init__()
        self.name = "Misaki VC"
        self.description = "Misaki voice conversion system"
        self.supported_languages = ["en", "ja"]  # Misaki primarily supports English and Japanese
        self.requirements = ["misaki"]
        self.logger = logging.getLogger(__name__)
    
    def is_available(self) -> bool:
        """Check if Misaki is available"""
        try:
            # Check if Misaki package is available
            import misaki
            return True
        except ImportError:
            return False
    
    async def clone_voice(self, source_audio: str, reference_audio: str, 
                         output_file: str, **kwargs) -> str:
        """Clone voice using Misaki"""
        if not self.is_available():
            raise RuntimeError("Misaki not available")
        
        # Placeholder implementation
        # In practice, this would use the Misaki voice conversion
        self.logger.info(f"Misaki voice cloning: {source_audio} -> {output_file}")
        
        # For now, just copy the source audio
        import shutil
        shutil.copy2(source_audio, output_file)
        
        return output_file
    
    async def train_voice_model(self, reference_audios: List[str], 
                               voice_name: str, **kwargs) -> str:
        """Train a voice model using Misaki"""
        if not self.is_available():
            raise RuntimeError("Misaki not available")
        
        # Placeholder implementation
        self.logger.info(f"Training Misaki voice model '{voice_name}' with {len(reference_audios)} files")
        
        return f"misaki_model_{voice_name}"