requirements.txt
# Core dependencies
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
soundfile>=0.10.0
pyyaml>=5.4.0
asyncio-throttle>=1.0.0

# Audio/Video processing
ffmpeg-python>=0.2.0

# STT Models
openai-whisper>=20231117
faster-whisper>=0.9.0
# whisperx  # Install separately if needed

# TTS Models  
bark>=1.0.0
piper-tts>=1.2.0
# kokoro  # Install separately if needed

# Voice Cloning
# Real-time voice cloning dependencies will be added based on specific implementation

# Lip Sync
opencv-python>=4.5.0
face-recognition>=1.3.0
dlib>=19.22.0

# Web framework (for API)
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# CLI
click>=8.0.0
rich>=10.0.0
tqdm>=4.62.0

# Utilities
requests>=2.25.0
pathlib2>=2.3.0
python-dotenv>=0.19.0

# Development
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Optional dependencies for specific models
# Uncomment as needed:

# For WhisperX
# whisperx>=3.1.0
# pyannote.audio>=3.0.0

# For Kokoro TTS  
# kokoro>=1.0.0

# For advanced voice cloning
# resemblyzer>=0.1.1
# webrtcvad>=2.0.10

# For GPU acceleration
# torch-audio>=0.9.0+cu111  # CUDA version
# torchaudio>=0.9.0+cu111   # CUDA version